+++ packages/dialog/src/dialog.tsx
@@ -24,22 +24,43 @@
  * IN THE SOFTWARE.
  */
 
-import { computed, defineComponent, onBeforeUnmount, onMounted, reactive, ref, watch } from 'vue';
+import { computed, defineComponent, reactive, ref } from 'vue';
 
 import Button from '@bkui-vue/button';
 import { useLocale, usePrefix } from '@bkui-vue/config-provider';
-import { Close, Error, Spinner, Success, Warn } from '@bkui-vue/icon';
+import { Error } from '@bkui-vue/icon';
 import Modal from '@bkui-vue/modal';
 
 import props from './props';
 
 export default defineComponent({
-  // eslint-disable-next-line vue/no-reserved-component-names
   name: 'Dialog',
   props,
-  emits: ['closed', 'update:isShow', 'confirm', 'prev', 'next', 'value-change'],
+  emits: {
+    closed: () => true,
+    shown: () => true,
+    hidden: () => true,
+    'update:isShow': (value: boolean) => value !== undefined,
+    confirm: () => true,
+    prev: () => true,
+    next: () => true,
+  },
   setup(props, { emit }) {
     const t = useLocale('dialog');
+
+    const { resolveClassName } = usePrefix();
+
+    const isMoveing = ref(false);
+
+    const data = reactive({
+      positionX: 0,
+      positionY: 0,
+      moveStyle: {
+        top: '50%',
+        left: '50%',
+      },
+    });
+
     const localConfirmText = computed(() => {
       if (props.confirmText === undefined) {
         return t.value.ok;
@@ -65,35 +86,6 @@ export default defineComponent({
       return props.nextText;
     });
 
-    const data = reactive({
-      positionX: 0,
-      positionY: 0,
-      moveStyle: {
-        top: '50%',
-        left: '50%',
-      },
-    });
-    const isModalShow = ref(props.isShow);
-
-    watch(
-      () => props.isShow,
-      (val: Boolean) => {
-        if (!val) {
-          setTimeout(() => {
-            data.moveStyle = {
-              top: '50%',
-              left: '50%',
-            };
-            data.positionX = 0;
-            data.positionY = 0;
-            isModalShow.value = false;
-          }, 250);
-        } else {
-          isModalShow.value = true;
-        }
-        emit('value-change', val);
-      },
-    );
     // 关闭弹框
     const handleClose = async () => {
       let shouldClose = true;
@@ -104,22 +96,22 @@ export default defineComponent({
       if (shouldClose) {
         emit('update:isShow', false);
         emit('closed');
-        isModalShow.value = false;
       }
     };
+
     const handleConfirm = () => {
       emit('update:isShow', false);
       emit('confirm');
     };
 
-    // 按 esc 关闭弹框
-    const escCloseHandler = e => {
-      if (props.isShow && props.closeIcon) {
-        if (e.keyCode === 27) {
-          handleClose();
-        }
-      }
+    const handleShown = () => {
+      emit('shown');
     };
+
+    const handleHidden = () => {
+      emit('hidden');
+    };
+
     // 上一步
     const handlePrevStep = () => {
       emit('prev');
@@ -130,7 +122,7 @@ export default defineComponent({
     };
 
     // 拖拽事件
-    const moveHandler = e => {
+    const handleMousedown = e => {
       if (props.fullscreen) {
         return false;
       }
@@ -149,6 +141,8 @@ export default defineComponent({
         disX = e.clientX - odiv.offsetLeft;
         disY = e.clientY - odiv.offsetTop;
       }
+      isMoveing.value = true;
+
       document.onmousemove = e => {
         const boxLeft = window.innerWidth - parentWidth;
         const boxTop = window.innerHeight - parentHeight;
@@ -169,69 +163,45 @@ export default defineComponent({
         data.moveStyle.left = `calc(50% + ${left}px)`;
         data.moveStyle.top = `calc(50% + ${top}px)`;
       };
+
       document.onmouseup = () => {
         document.onmousemove = null;
         document.onmouseup = null;
+        isMoveing.value = false;
       };
     };
 
-    const { resolveClassName } = usePrefix();
-
-    onMounted(() => {
-      if (props.escClose) {
-        addEventListener('keydown', escCloseHandler);
-      }
-    });
-    onBeforeUnmount(() => {
-      if (props.escClose) {
-        removeEventListener('keydown', escCloseHandler);
-      }
-    });
-
     return {
       data,
-      handleClose,
-      handleConfirm,
-      escCloseHandler,
-      moveHandler,
-      handlePrevStep,
-      handleNextStep,
-      isModalShow,
       localConfirmText,
       localCancelText,
       localPrevText,
       localNextText,
       resolveClassName,
+      handleHidden,
+      handleShown,
+      handleClose,
+      handleConfirm,
+      handleMousedown,
+      handlePrevStep,
+      handleNextStep,
     };
   },
 
   render() {
-    const renderIcon = () => {
-      const iconMap = {
-        loading: <Spinner class={[this.resolveClassName('info-icon'), 'primary']}></Spinner>,
-        warning: <Warn class={[this.resolveClassName('info-icon'), 'warning']}></Warn>,
-        success: <Success class={[this.resolveClassName('info-icon'), 'success']}></Success>,
-        danger: <Close class={[this.resolveClassName('info-icon'), 'danger']}></Close>,
-      };
-      return iconMap[this.infoType];
-    };
-
     const dialogSlot = {
       header: () => [
         <div
-          class={[
-            this.resolveClassName('dialog-tool'),
-            this.fullscreen || !this.draggable ? '' : 'move',
-            this.draggable ? 'content-dragging' : '',
-          ]}
-          onMousedown={this.moveHandler}
+          class={{
+            [this.resolveClassName('dialog-tool')]: true,
+            'is-draggable': !this.fullscreen && this.draggable,
+            'is-dragging': this.draggable,
+          }}
+          onMousedown={this.handleMousedown}
         >
-          {this.$slots.tools?.() ?? ''}
+          {this.$slots.tools?.()}
         </div>,
         <div class={this.resolveClassName('dialog-header')}>
-          <div class={this.resolveClassName('dialog-header-icon')}>
-            {this.infoType ? renderIcon() : <slot name='info-icon' />}
-          </div>
           <span
             class={this.resolveClassName('dialog-title')}
             style={`text-align: ${this.headerAlign}`}
@@ -262,7 +232,7 @@ export default defineComponent({
               <>
                 <Button
                   onClick={this.handleConfirm}
-                  theme={this.theme}
+                  theme={this.confirmButtonTheme}
                   loading={this.isLoading}
                 >
                   {this.localConfirmText}
@@ -281,7 +251,7 @@ export default defineComponent({
             return (
               <Button
                 onClick={this.handleConfirm}
-                theme={this.theme}
+                theme={this.confirmButtonTheme}
                 loading={this.isLoading}
               >
                 {this.localConfirmText}
@@ -313,7 +283,7 @@ export default defineComponent({
                     </Button>
                     <Button
                       onClick={this.handleConfirm}
-                      theme={this.theme}
+                      theme={this.confirmButtonTheme}
                       loading={this.isLoading}
                     >
                       {this.localConfirmText}
@@ -347,21 +317,32 @@ export default defineComponent({
           </div>
         );
       },
-      close: () => <Error onClick={this.handleClose} />,
+      close: () => <Error />,
     };
 
     return (
       <Modal
-        {...this.$props}
         class={{
           [this.resolveClassName('dialog')]: true,
           [this.resolveClassName('dialog-wrapper')]: true,
           'is-fullscreen': this.fullscreen,
         }}
+        isShow={this.isShow}
+        fullscreen={this.fullscreen}
+        width={this.fullscreen ? 'auto' : this.width}
+        animateType='fadein'
+        beforeClose={this.beforeClose}
+        closeIcon={this.closeIcon}
+        escClose={this.escClose}
+        quickClose={this.quickClose}
+        showMask={this.showMask}
+        transfer={this.transfer}
+        left={this.fullscreen ? '0px' : this.data.moveStyle.left}
+        top={this.fullscreen ? '0px' : this.data.moveStyle.top}
+        zIndex={this.zIndex}
         onClose={this.handleClose}
-        isShow={this.isModalShow}
-        left={this.data.moveStyle.left}
-        top={this.data.moveStyle.top}
+        onHidden={this.handleHidden}
+        onShown={this.handleShown}
       >
         {dialogSlot}
       </Modal>