+++ packages/modal/src/hooks.ts
@@ -24,33 +24,28 @@
  * IN THE SOFTWARE.
  */
 import throttle from 'lodash/throttle';
-import { type Ref, ref, useSlots, watch } from 'vue';
+import { type Ref, ref, watch } from 'vue';
 
 import { usePrefix } from '@bkui-vue/config-provider';
 
 import type { ModalProps } from './modal';
 
-export const useContentResize = (root: Ref<HTMLElement>, props: ModalProps) => {
+export const useContentResize = (root: Ref<HTMLElement>, resizeTarget: Ref<HTMLElement>, props: ModalProps) => {
   const { resolveClassName } = usePrefix();
-  const slots = useSlots();
 
   const isContentScroll = ref(false);
   const contentStyles = ref({});
 
-  let observer;
+  let observer: ResizeObserver;
 
   const handleContentBoxChange = () => {
-    if (!slots.footer) {
-      return;
-    }
-
     const calcContentScroll = throttle(() => {
       const { height: headerHeight } = root.value
         .querySelector(`.${resolveClassName('modal-header')}`)
         .getBoundingClientRect();
 
       const { height: contentHeight } = root.value
-        .querySelector(`.${resolveClassName('modal-content')}`)
+        .querySelector(`.${resolveClassName('modal-content')} div`)
         .getBoundingClientRect();
 
       const { height: footerHeight } = root.value
@@ -59,26 +54,25 @@ export const useContentResize = (root: Ref<HTMLElement>, props: ModalProps) => {
 
       const windowInnerHeight = window.innerHeight;
 
-      isContentScroll.value = windowInnerHeight < headerHeight + contentHeight + footerHeight;
+      isContentScroll.value = windowInnerHeight < headerHeight + contentHeight + footerHeight + 20;
       if (isContentScroll.value || props.fullscreen) {
         contentStyles.value = {
           height: `${windowInnerHeight - headerHeight - footerHeight}px`,
-          overflow: 'scroll',
+          overflow: 'auto',
+          'scrollbar-gutter': 'stable',
         };
+        // fullscreen 时默认为 true
+        isContentScroll.value = true;
       } else {
         contentStyles.value = {};
       }
-    }, 100);
+    }, 30);
 
-    observer = new MutationObserver(() => {
+    observer = new ResizeObserver(() => {
       calcContentScroll();
     });
 
-    observer.observe(root.value.querySelector(`.${resolveClassName('modal-content')} div`), {
-      subtree: true,
-      attributes: true,
-      childList: true,
-    });
+    observer.observe(resizeTarget.value);
 
     calcContentScroll();
   };
@@ -89,10 +83,9 @@ export const useContentResize = (root: Ref<HTMLElement>, props: ModalProps) => {
       if (props.isShow) {
         setTimeout(() => {
           handleContentBoxChange();
-        }, 100);
+        });
       } else {
         if (observer) {
-          observer.takeRecords();
           observer.disconnect();
           observer = null;
         }