+++ site/views/virtual-render/basic.vue
@@ -14,7 +14,6 @@
     </div>
     <bk-virtual-render
       ref="refFixToTop"
-      :scrollbar="scrollbar"
       :list="randomRows"
       :line-height="30"
       :height="300"
@@ -114,7 +113,7 @@
         this.randomRows.splice(
           0,
           this.randomRows.length,
-          ...new Array(Math.ceil(Math.random() * 900) + 100).fill('')
+          ...new Array(Math.ceil(Math.random() * 9000) + 1000).fill('')
             .map((_, index) => ({
               ip: `${index}--192.168.0.x`,
               source: `${index}_QQ`,